vilanova/imagensIonv/requirements.txt
cx_Oracle==8.3.0
python-dotenv==1.0.1
pytz==2024.2
